feat: use new service to connect to lcms instead of airtable

+++ src/learning-content.js
@@ -96,12 +96,14 @@ const tables = [{
 }];
 
 async function fetchAndSaveData(configuration) {
-  const learningContent = lcms.getLearningContent(configuration);
-  await Promise.all(tables.map(async (table) => {
-    await dbConnection.dropTable(table.name, configuration);
-    await dbConnection.createTable(table, configuration);
-    await dbConnection.saveLearningContent(table, learningContent[table.name], configuration);
-  }));
+  const learningContent = await lcms.getLearningContent(configuration);
+  if (learningContent) {
+    await Promise.all(tables.map(async (table) => {
+      await dbConnection.dropTable(table.name, configuration);
+      await dbConnection.createTable(table, configuration);
+      await dbConnection.saveLearningContent(table, learningContent[table.name], configuration);
+    }));
+  }
 }
 
 module.exports = {
